pyproject.toml
[tool.poetry]
name = "CandleSticks"
version = "0.1.0"
description = ""
authors = ["leytzher <leytzher@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
scipy = "^1.7.0"
pyts = "^0.11.0"
pandas = "^1.3.1"
python-binance = "^1.0.12"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
